pyproject.toml
[tool.poetry]
name = "tsw-cli"
version = "0.1.0"
description = "tsw-cli"
authors = ["foxgem <jianhgreat@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.15.1"
agno = "^1.0.3"
markdown-pdf = "^1.3.3"
pydantic-settings = "^2.7.1"
pydantic-core = "^2.28.0"
pydantic = "^2.10.6"
google-generativeai = "^0.8.4"
python-dotenv = "^1.0.1"
duckduckgo-search = "^7.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
